Handle twilight events concurrently to avoid deadlocks
Songbird requires that Songbird::join is not awaited in the same task
as the stream of Twilight events, because Songbird::join will never
resolve if the events are not being passed from somewhere else via
Songbird::process.

The suggested way of handling this is to spawn app event handling after
calling Songbird::process and NOT await it. This works however it makes
fatal error handling really annoying because the errors have to be
passed via a channel.
Instead, it's more straightforward to handle the twilight events
concurrently AND in parallel, by spawning a task for each event and
awaiting for potential failures via try_for_each_concurrently Stream
utility. In case of a failure the stream will exit early with that
failure, in case of no failures it will continue running forever.